# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/integrations/pricing-agents/{agentId}/calculate": {
      "post": {
        "operationId": "IntegrationsController_calculatePrice",
        "parameters": [
          {
            "name": "agentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "API token (Bearer token)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Price calculated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteResultDTO"
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteResultDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid parameters or agent not deployed"
          },
          "404": {
            "description": "Pricing agent or checkpoint not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Calculate price using agent schema parameters",
        "tags": [
          "integrations"
        ]
      }
    },
    "/integrations/pricing-agents/{agentId}/chat": {
      "post": {
        "operationId": "IntegrationsController_processConversation",
        "parameters": [
          {
            "name": "agentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "API token (Bearer token)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaygroundExecutionRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Conversation processed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaygroundExecutionResponseDto"
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaygroundExecutionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid input or agent not deployed"
          },
          "404": {
            "description": "Pricing agent or checkpoint not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Process human language conversation for pricing calculation",
        "tags": [
          "integrations"
        ]
      }
    },
...
  },
  "info": {
    "title": "Pricing Agent Builder API",
    "description": "API documentation for the Pricing Agent Builder service",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "ChatbotMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "The message text"
          },
          "role": {
            "type": "string",
            "description": "The role of the message sender",
            "enum": [
              "AI",
              "User"
            ]
          }
        },
        "required": [
          "message",
          "role"
        ]
      },
      "PlaygroundExecutionRequestDto": {
        "type": "object",
        "properties": {
          "input": {
            "type": "string"
          },
          "conversation": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatbotMessage"
            }
          }
        },
        "required": [
          "input"
        ]
      },
      "BacktraceCalculationStepDTO": {
        "type": "object",
        "properties": {
          "operation": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "subTasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BacktraceCalculationStepDTO"
            }
          }
        },
        "required": [
          "operation",
          "description"
        ]
      },
      "QuoteErrorDTO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "QuoteResultDTO": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "pricingCalculationBacktrace": {
            "$ref": "#/components/schemas/BacktraceCalculationStepDTO"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuoteErrorDTO"
            }
          }
        }
      },
      "PlaygroundExecutionResponseDto": {
        "type": "object",
        "properties": {
          "structuredOrder": {
            "type": "object",
            "description": "The structured order input converted from natural language"
          },
          "functionResult": {
            "description": "The result from executing the pricing function",
            "allOf": [
              {
                "$ref": "#/components/schemas/QuoteResultDTO"
              }
            ]
          },
          "aiMessage": {
            "type": "string",
            "description": "The AI-generated response message for the playground conversation"
          }
        },
        "required": [
          "structuredOrder",
          "functionResult",
          "aiMessage"
        ]
      },
...
    }
  }
}